+++ CHANGELOG.md
@@ -44,6 +44,9 @@
 ### Formatter
 
 - The format used by the formatter has been improved in some niche cases.
+- Improved the formatting of long case guards.
+- The formatter can now format groups of imports alphabetically.
+
 
 ## v1.0.0-rc1 - 2024-02-10